+++ src/Microsoft.Health.Fhir.Core/Features/Operations/ConvertData/ITemplateProviderFactory.cs
@@ -1,0 +1,14 @@
+﻿// -------------------------------------------------------------------------------------------------
+// Copyright (c) Microsoft Corporation. All rights reserved.
+// Licensed under the MIT License (MIT). See LICENSE in the repo root for license information.
+// -------------------------------------------------------------------------------------------------
+
+namespace Microsoft.Health.Fhir.Core.Features.Operations.ConvertData
+{
+    public interface ITemplateProviderFactory
+    {
+        public IConvertDataTemplateProvider GetContainerRegistryTemplateProvider();
+
+        public IConvertDataTemplateProvider GetDefaultTemplateProvider();
+    }
+}
