+++ src/Microsoft.Health.Fhir.Shared.Core.UnitTests/Features/Operations/ConvertData/ContainerRegistryTemplateProviderTests.cs
@@ -26,9 +26,9 @@ namespace Microsoft.Health.Fhir.Shared.Core.UnitTests.Features.Operations.Conver
     public class ContainerRegistryTemplateProviderTests
     {
         [Fact]
-        public async Task GivenDefaultTemplateReference_ForAcrTemplateProvider_WhenFetchingTemplates_DefaultTemplateCollectionShouldReturn()
+        public async Task GivenDefaultTemplateReference_WhenFetchingTemplates_DefaultTemplateCollectionShouldReturn()
         {
-            var containerRegistryTemplateProvider = GetContainerRegistryTemplateProvider();
+            var containerRegistryTemplateProvider = GetDefaultTemplateProvider();
             foreach (var templateInfo in DefaultTemplateInfo.DefaultTemplateMap.Values)
             {
                 var templateReference = templateInfo.ImageReference;
@@ -41,13 +41,26 @@ namespace Microsoft.Health.Fhir.Shared.Core.UnitTests.Features.Operations.Conver
         [Fact]
         public async Task GivenAnInvalidToken_WhenFetchingCustomTemplates_ExceptionShouldBeThrown()
         {
-            var containerRegistryTemplateProvider = GetContainerRegistryTemplateProvider();
+            var containerRegistryTemplateProvider = GetCustomTemplateProvider();
             var templateReference = "test.azurecr.io/templates:latest";
 
             await Assert.ThrowsAsync<ContainerRegistryNotAuthorizedException>(() => containerRegistryTemplateProvider.GetTemplateCollectionAsync(GetRequestWithTemplateReference(templateReference), CancellationToken.None));
         }
 
-        private IConvertDataTemplateProvider GetContainerRegistryTemplateProvider()
+        private IConvertDataTemplateProvider GetDefaultTemplateProvider()
+        {
+            var convertDataConfig = new ConvertDataConfiguration
+            {
+                Enabled = true,
+                OperationTimeout = TimeSpan.FromSeconds(1),
+            };
+            convertDataConfig.ContainerRegistryServers.Add("test.azurecr.io");
+
+            var config = Options.Create(convertDataConfig);
+            return new DefaultTemplateProvider(config, new NullLogger<ContainerRegistryTemplateProvider>());
+        }
+
+        private IConvertDataTemplateProvider GetCustomTemplateProvider()
         {
             IContainerRegistryTokenProvider tokenProvider = Substitute.For<IContainerRegistryTokenProvider>();
             tokenProvider.GetTokenAsync(default, default).ReturnsForAnyArgs("Bearer faketoken");
