+++ src/Microsoft.Health.Fhir.Shared.Web/appsettings.json
@@ -1,7 +1,7 @@
 ﻿{
     "FhirServer": {
         "Security": {
-            "Enabled": true,
+            "Enabled": false,
             "EnableAadSmartOnFhirProxy": true,
             "Authentication": {
                 "Audience": null,
@@ -46,9 +46,16 @@
             "AllowCredentials": false
         },
         "Operations": {
-            "HostingBackgroundServiceQueues":
-                [{ "Queue": "Import", "UpdateProgressOnHeartbeat": true },
-                 { "Queue": "Export", "UpdateProgressOnHeartbeat": false }],
+            "HostingBackgroundServiceQueues": [
+                {
+                    "Queue": "Import",
+                    "UpdateProgressOnHeartbeat": true
+                },
+                {
+                    "Queue": "Export",
+                    "UpdateProgressOnHeartbeat": false
+                }
+            ],
             "Export": {
                 "Enabled": true,
                 "StorageAccountConnection": null,
@@ -75,7 +82,7 @@
                 "QueryDelayIntervalInMilliseconds": 500
             },
             "ConvertData": {
-                "Enabled": false,
+                "Enabled": true,
                 "ContainerRegistryServers": [],
                 "TemplateCollectionOptions": {
                     "ShortCacheTimeSpan": "00:30:00",
@@ -117,9 +124,9 @@
         "ArtifactStore": {
             "OciArtifacts": [
                 {
-                    "LoginServer": "",
-                    "ImageName": "",
-                    "Digest": ""
+                    "LoginServer": "telifhirtemplates.azurecr.io",
+                    "ImageName": "templatetest",
+                    "Digest": "sha256:f0f8552406d3e775266d84a3a5b47f817aeeadd8a442d47a16e65692b7f770bf"
                 }
             ]
         },
@@ -159,9 +166,22 @@
             "EnableConcurrencyIfQueryExceedsTimeLimit": true
         }
     },
-    "DataStore": "CosmosDb",
+    "DataStore": "SqlServer",
     "SqlServer": {
-        "CommandTimeout": "00:03:00"
+        "CommandTimeout": "00:03:00",
+        "Initialize": "true",
+        "AllowDatabaseCreation": "true",
+        "AuthenticationType": "ConnectionString",
+        "TransientFaultRetryPolicy": {
+            "InitialDelay": "00:00:00.100",
+            "RetryCount": 3,
+            "Factor": 2,
+            "FastFirst": true
+        },
+        "SchemaOptions": {
+            "AutomaticUpdatesEnabled": true
+        },
+        "ConnectionString": "Server=tcp:127.0.0.1,1433;Initial Catalog=l9uxxrqa6elqiqo3u2kofi7i;TrustServerCertificate=True;Connection Timeout=30;User Id=SA;Password=Shaund99;"
     },
     "KeyVault": {
         "Endpoint": null
